server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Server"
authors = ["Javin Lacson <javin@teamengine.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
marshmallow = "^3.19.0"
djangorestframework = "3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
